authorization for reviews and listing

+++ 47. Major Project/controllers/listingRoutes.js
@@ -1,11 +1,14 @@
+const jwt = require("jsonwebtoken");
 const Listing = require("../models/listing");
 const wrapAsync = require("../utils/wrapAsync.js");
 const ExpressError = require("../utils/ExpressErrors");
-
 const listingRouter = require("express").Router();
 const { listingSchema } = require("../Schema.js");
-
-const checkAndHanldeNotLoggedIn = require("../utils/module.js");
+require("dotenv").config();
+const {
+  checkAndHanldeNotLoggedIn,
+  checkIfAuthorized,
+} = require("../utils/module.js");
 
 const validateListing = (req, res, next) => {
   const result = listingSchema.validate({ listing: req.body });
@@ -18,25 +21,71 @@ listingRouter.get(
   "/",
   wrapAsync(async (req, res) => {
     const allListings = await Listing.find({});
-    res.render("listings/index", { allListings });
+    if (res.locals.token) {
+      const decodedToken = jwt.verify(res.locals.token, process.env.JWT_SECRET);
+      const currUserId = decodedToken.id;
+      res.render("listings/index", {
+        allListings: allListings.map((i) => {
+          if (i.owner == currUserId) i["isOwner"] = true;
+          return i;
+        }),
+      });
+    } else {
+      res.render("listings/index", { allListings });
+    }
   })
 );
 
 listingRouter.get(
   "/:id",
   wrapAsync(async (req, res) => {
-    const listing = await Listing.findById(req.params.id).populate("reviews");
+    let listing = await Listing.findById(req.params.id).populate([
+      "reviews",
+      "owner",
+    ]);
+    if (res.locals.token) {
+      const decodedToken = jwt.verify(res.locals.token, process.env.JWT_SECRET);
+      const currUserId = decodedToken.id;
+      listing = {
+        _id: listing._id,
+        title: listing.title,
+        description: listing.description,
+        image: listing.image,
+        price: listing.price,
+        location: listing.location,
+        country: listing.country,
+        owner: listing.owner,
+        reviews: listing.reviews.map((i) => {
+          if (i.owner == currUserId)
+            i.isOwner = true;
+          return i;
+        }),
+      };     
+    }
+
+    res.locals.isAuthorizedToEdit = false;
+    if (res.locals.token) {
+      checkIfAuthorized(res.locals.token, res, listing);
+    }
     if (!listing) {
       req.flash("error", `Listing not found`);
       res.redirect("/");
     } else {
-      let netRating = listing.reviews.map((i) => i.rating);
-      netRating = (
-        netRating.reduce((accumulator, currentValue) => {
-          return accumulator + currentValue;
-        }, 0) / netRating.length
-      ).toFixed(1);
-      res.render("listings/place", { place: listing, netRating });
+      if (listing.reviews.length > 0) {
+        let netRating = listing.reviews.map((i) => i.rating);
+        netRating = (
+          netRating.reduce((accumulator, currentValue) => {
+            return accumulator + currentValue;
+          }, 0) / netRating.length
+        ).toFixed(1);
+        if (!res.locals.netRating) res.locals.netRating = "No Ratings";
+        res.render("listings/place", { place: listing, netRating });
+      } else {
+        res.render("listings/place", {
+          place: listing,
+          netRating: "No Ratings",
+        });
+      }
     }
   })
 );
@@ -45,27 +94,37 @@ listingRouter.post(
   "/",
   // validateListing,
   wrapAsync(async (req, res, next) => {
-    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to add a new listing`)){
-      return;
+    if (res.locals.token) {
+      const decodedToken = jwt.verify(res.locals.token, process.env.JWT_SECRET);
+      const currUserId = decodedToken.id;
+      const newListing = new Listing({ ...req.body, owner: currUserId });
+      await newListing
+        .save()
+        .then((res) => {
+          req.flash("success", `'${res.title}' Listing Created Successfully`);
+          res.redirect(`/${res._id}`);
+        })
+        .catch((err) => {
+          req.flash("error", err);
+          res.redirect(`/`);
+        });
+    } else {
+      req.flash("error", `You must be logged in to add a new listing`);
+      res.redirect("/?startlogin=true");
     }
-    // if(!req.body.title && !req.body.description && !req.body.image && !req.body.price && !req.body.location && !req.body.country){}
-    const newListing = new Listing(req.body);
-    await newListing
-      .save()
-      .then((res) => {
-        req.flash("success", `'${res.title}' Listing Created Successfully`);
-      })
-      .catch((err) => {
-        req.flash("error", err);
-      });
-    res.redirect("/");
   })
 );
 
 listingRouter.get(
   "/edit/:id",
   wrapAsync(async (req, res) => {
-    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to edit a listing`)){
+    if (
+      checkAndHanldeNotLoggedIn(
+        req,
+        res,
+        `You must be logged in to edit a listing`
+      )
+    ) {
       return;
     }
     const listing = await Listing.findById(req.params.id);
@@ -73,7 +132,19 @@ listingRouter.get(
       req.flash("error", `Listing not found`);
       res.redirect("/");
     } else {
-      res.render("listings/edit", { place: listing });
+      if (
+        res.locals.token &&
+        checkIfAuthorized(
+          res.locals.token,
+          res,
+          await Listing.findById(req.params.id)
+        )
+      ) {
+        res.render("listings/edit", { place: listing });
+      } else {
+        req.flash("error", `You are not authorized to edit this listing`);
+        res.redirect(`/${req.params.id}`);
+      }
     }
   })
 );
@@ -82,16 +153,33 @@ listingRouter.patch(
   "/:id",
   // validateListing,
   wrapAsync(async (req, res) => {
-    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to edit a listing`)){
+    if (
+      checkAndHanldeNotLoggedIn(
+        req,
+        res,
+        `You must be logged in to edit a listing`
+      )
+    ) {
       return;
     }
-    await Listing.findByIdAndUpdate(req.params.id, { $set: req.body })
-      .then((res) => {
-        req.flash("success", `'${res.title}' Listing Updated Successfully`);
-      })
-      .catch((err) => {
-        req.flash("error", err);
-      });
+    if (
+      res.locals.token &&
+      checkIfAuthorized(
+        res.locals.token,
+        res,
+        await Listing.findById(req.params.id)
+      )
+    ) {
+      await Listing.findByIdAndUpdate(req.params.id, { $set: req.body })
+        .then((res) => {
+          req.flash("success", `'${res.title}' Listing Updated Successfully`);
+        })
+        .catch((err) => {
+          req.flash("error", err);
+        });
+    } else {
+      req.flash("error", `You are not authorized to edit this listing`);
+    }
     res.redirect(`/${req.params.id}`);
   })
 );
@@ -99,18 +187,36 @@ listingRouter.patch(
 listingRouter.delete(
   "/:id",
   wrapAsync(async (req, res) => {
-    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to delete a listing`)){
+    if (
+      checkAndHanldeNotLoggedIn(
+        req,
+        res,
+        `You must be logged in to delete a listing`
+      )
+    ) {
       return;
     }
-    await Listing.findByIdAndDelete(req.params.id)
-      .then((res) => {
-        req.flash("success", `'${res.title}' Listing Deleted Successfully`);
-      })
-      .catch((err) => {
-        req.flash("error", err);
-      });
-    // await Review.deleteMany({listing:req.params.id})
-    // instead of the above action we define a post mongoose middleware (check it in models/listing.js)
+    if (
+      res.locals.token &&
+      checkIfAuthorized(
+        res.locals.token,
+        res,
+        await Listing.findById(req.params.id)
+      )
+    ) {
+      await Listing.findByIdAndDelete(req.params.id)
+        .then((res) => {
+          req.flash("success", `'${res.title}' Listing Deleted Successfully`);
+        })
+        .catch((err) => {
+          req.flash("error", err);
+        });
+      // await Review.deleteMany({listing:req.params.id})
+      // instead of the above action we define a post mongoose middleware (check it in models/listing.js)
+    } else {
+      req.flash("error", `You are not authorized to delete this listing`);
+      res.redirect(`/${req.params.id}`);
+    }
     res.redirect("/");
   })
 );
@@ -118,7 +224,13 @@ listingRouter.delete(
 listingRouter.get(
   "/addNew/new",
   wrapAsync(async (req, res) => {
-    if(checkAndHanldeNotLoggedIn(req,res,`You must be logged in to add a new listing`)){
+    if (
+      checkAndHanldeNotLoggedIn(
+        req,
+        res,
+        `You must be logged in to add a new listing`
+      )
+    ) {
       return;
     }
     res.render("listings/newListing");
